refactor(vscode): lift updateCardMetadata + applyRelativeSizing into cardBuilder
Continues the card-lifecycle extraction from #814. Lifts the remaining
two metadata-side functions out of `behavior.ts` into `cardBuilder.ts`,
where they sit alongside `buildPreviewCard`. The eventual
`<preview-card>` Lit component will fold all of these (build, update,
relative-sizing, plus the still-imperative `updateImage` and
`applyA11yUpdate`) into one reactive `render()`.

New `cardBuilder.ts` exports:
- `updateCardMetadata(card, p, config)` — refresh dataset, title, capture
  cache (preserving in-flight imageData), variant badge, and a11y
  legend / overlay. Takes a narrow `CardUpdateConfig` (a `Pick` of
  `CardBuilderConfig` covering `cardCaptures`, `frameCarousel`,
  `earlyFeatures`).
- `applyRelativeSizing(previews)` — pure DOM walk over all cards in
  the grid, sets `--size-ratio` / `--aspect-ratio` CSS vars from each
  preview's `widthDp` / `heightDp`. No config needed.

`behavior.ts` calls the new functions from `renderPreviews` (existing
card path) and `setPreviews` (post-render relative-sizing pass), with
the per-call config built once at panel boot. Same DOM, same
side effects. No behaviour change.

## Drive-by

- Drops the now-unused `buildTooltip`, `buildVariantLabel`,
  `isAnimatedPreview`, `isWearPreview` imports from `behavior.ts` —
  all four were only referenced from `updateCardMetadata` and lifted
  along with it.
- Replaces a stale comment with a real type annotation on
  `cardCaptures`: was `// Map<previewId, CapturePresentation[]>`
  followed by `new Map()`; now `new Map<string, CapturePresentation[]>()`.
- Updates the `cardBuilder.ts` header comment to reflect the wider
  scope (lifecycle, not just initial build).

`behavior.ts`: 1141 → 1032 lines (−109 LOC).
`cardBuilder.ts`: 224 → 361 lines (+137 LOC).

## Test plan

- [x] `npm run compile` clean (host + webview)
- [x] `npm test` — 350/350 passing
- [x] `npm run format:check` clean
- [ ] Manual smoke: open the panel, edit a `@Preview` to trigger a fresh
  `setPreviews` for an existing previewId. Verify:
  - Title updates if `functionName` / `params.name` changed.
  - Variant badge appears / updates / disappears for changing
    `device` / `widthDp` / `heightDp` / `fontScale` / `uiMode` / `locale`.
  - In-flight imageData survives a metadata-only refresh.
  - Capture-count change clamps `currentIndex`.
  - Toggle `composePreview.earlyFeatures` mid-session — a11y legend
    + overlay appear / disappear without a full re-render.
  - Mixed widthDp/heightDp cards size proportionally in `flow` /
    `column` layouts.
