+++ vscode-extension/src/webview/preview/cardBuilder.ts
@@ -1,21 +1,17 @@
-// Initial DOM build for a single `<div class="preview-card">`.
+// Lifecycle DOM operations for a single `<div class="preview-card">`:
+// initial build (`buildPreviewCard`), metadata refresh on a fresh
+// `setPreviews` (`updateCardMetadata`), and the grid-wide relative-sizing
+// pass (`applyRelativeSizing`).
 //
-// Lifted verbatim from `behavior.ts`'s `createCard` so the imperative DOM
-// construction stops needing closure access to the rest of the panel. The
-// dynamic update paths (`updateCardMetadata`, `updateImage`, `applyA11yUpdate`)
-// still live in `behavior.ts` for now — they query the card via
-// `document.getElementById("preview-" + sanitizeId(id))` and patch in place,
-// independent of the initial build path.
-//
-// The eventual `<preview-card>` Lit component will fold both the initial
-// build and the dynamic updates into one reactive `render()`. This module is
-// the bridge: it captures every closed-over collaborator
-// (`vscode`, `liveState`, `staleBadge`, `frameCarousel`, viewport observer,
-// the `cardCaptures` map, the `interactiveInputConfig`) into a typed config
-// so the same logic can move into the component without re-resolving each
-// dependency through `setupPreviewBehavior`.
+// Lifted verbatim from `behavior.ts`'s `createCard` / `updateCardMetadata` /
+// `applyRelativeSizing` so the imperative DOM operations stop needing
+// closure access to the rest of the panel. The remaining update paths —
+// `updateImage` (per-frame image swap) and `applyA11yUpdate` (daemon-
+// attached a11y refresh) — still live in `behavior.ts` and patch the card
+// in place via `document.getElementById`. The eventual `<preview-card>`
+// Lit component will fold all five into a single reactive `render()`.
 
-import { buildA11yLegend } from "./a11yOverlay";
+import { buildA11yLegend, buildA11yOverlay } from "./a11yOverlay";
 import {
     buildTooltip,
     buildVariantLabel,
@@ -225,4 +221,141 @@ export function buildPreviewCard(
 
     config.observeForViewport(card);
     return card;
+}
+
+/** Subset of `CardBuilderConfig` that `updateCardMetadata` actually
+ *  reaches for. Kept narrow so callers — and the eventual reactive
+ *  Lit component — only have to satisfy what's used. */
+export type CardUpdateConfig = Pick<
+    CardBuilderConfig,
+    "cardCaptures" | "frameCarousel" | "earlyFeatures"
+>;
+
+/**
+ * Refresh an existing card after a `setPreviews` for an id we already
+ * have in the grid. Patches the card's dataset, title text, capture cache
+ * (preserving already-received `imageData` for surviving renderOutputs),
+ * variant badge, and a11y legend / overlay layer.
+ *
+ * Caller is responsible for finding the card — `renderPreviews` walks the
+ * grid once and dispatches to `updateCardMetadata` vs `buildPreviewCard`
+ * based on whether the previewId existed previously.
+ */
+export function updateCardMetadata(
+    card: HTMLElement,
+    p: PreviewInfo,
+    config: CardUpdateConfig,
+): void {
+    card.dataset.function = p.functionName;
+    card.dataset.group = p.params.group || "";
+    card.dataset.wearPreview = isWearPreview(p) ? "1" : "0";
+    const title = card.querySelector<HTMLButtonElement>(".card-title");
+    if (title) {
+        title.textContent =
+            p.functionName + (p.params.name ? " — " + p.params.name : "");
+        title.title = buildTooltip(p);
+    }
+    // Refresh capture labels in place. If the capture count changed
+    // (e.g. user edited @RoboComposePreviewOptions) we preserve
+    // already-received imageData for renderOutputs that carry over.
+    const newCaps = p.captures.map((c) => ({
+        renderOutput: c.renderOutput,
+        label: c.label || "",
+    }));
+    const prior = config.cardCaptures.get(p.id) ?? [];
+    // Match by index rather than renderOutput since filenames may
+    // legitimately change (e.g. a preview gains a @RoboComposePreviewOptions
+    // annotation). Mismatched positions just reset to null-image.
+    const mergedCaps = newCaps.map(
+        (nc, i): CapturePresentation => ({
+            label: nc.label,
+            renderOutput: nc.renderOutput || "",
+            imageData: prior[i]?.imageData ?? null,
+            errorMessage: prior[i]?.errorMessage ?? null,
+            renderError: prior[i]?.renderError ?? null,
+        }),
+    );
+    config.cardCaptures.set(p.id, mergedCaps);
+    const curIdx = parseInt(card.dataset.currentIndex || "0", 10);
+    if (curIdx >= mergedCaps.length) {
+        card.dataset.currentIndex = String(Math.max(0, mergedCaps.length - 1));
+    }
+    if (isAnimatedPreview(p)) config.frameCarousel.updateIndicator(card);
+    const variantLabel = buildVariantLabel(p);
+    let badge = card.querySelector(".variant-badge");
+    if (variantLabel) {
+        if (!badge) {
+            badge = document.createElement("div");
+            badge.className = "variant-badge";
+            card.appendChild(badge);
+        }
+        badge.textContent = variantLabel;
+    } else if (badge) {
+        badge.remove();
+    }
+
+    // Refresh the a11y legend + overlay in place when findings
+    // change (e.g. toggling a11y on turns findings from null → list,
+    // or a fresh render updates the set). Tear down the old nodes
+    // and rebuild: simpler than reconciling row-by-row for what is
+    // a rare event.
+    const existingLegend = card.querySelector(".a11y-legend");
+    const existingOverlay = card.querySelector(".a11y-overlay");
+    if (existingLegend) existingLegend.remove();
+    if (existingOverlay) existingOverlay.innerHTML = "";
+    if (config.earlyFeatures() && p.a11yFindings && p.a11yFindings.length > 0) {
+        const container = card.querySelector(".image-container");
+        if (container && !container.querySelector(".a11y-overlay")) {
+            const overlay = document.createElement("div");
+            overlay.className = "a11y-overlay";
+            overlay.setAttribute("aria-hidden", "true");
+            container.appendChild(overlay);
+        }
+        const legend = buildA11yLegend(card, p);
+        card.appendChild(legend);
+        // Repopulate box geometry if the image is already loaded —
+        // otherwise updateImage's load handler will pick it up on
+        // the next render cycle.
+        const img = card.querySelector<HTMLImageElement>(
+            ".image-container img",
+        );
+        if (img && img.complete && img.naturalWidth > 0) {
+            buildA11yOverlay(card, p.a11yFindings, img);
+        }
+    } else if (existingOverlay) {
+        // No findings or feature off — drop any leftover overlay
+        // div so cards stay clean when the user toggles
+        // earlyFeatures off mid-session.
+        existingOverlay.remove();
+    }
+}
+
+/**
+ * Scale image containers so preview variants at different device sizes
+ * (e.g. wearos_large_round 227dp vs wearos_small_round 192dp) render at
+ * relative sizes in fixed-layout modes. Only applied when we have real
+ * widthDp/heightDp — variants without known dimensions fall back to the
+ * default CSS (full card width, auto aspect).
+ *
+ * Walks all of [previews] once; idempotent — subsequent calls overwrite
+ * the CSS custom properties cleanly.
+ */
+export function applyRelativeSizing(previews: readonly PreviewInfo[]): void {
+    const widths = previews
+        .map((p) => p.params.widthDp ?? 0)
+        .filter((w) => w > 0);
+    const maxW = widths.length > 0 ? Math.max(...widths) : 0;
+    for (const p of previews) {
+        const card = document.getElementById("preview-" + sanitizeId(p.id));
+        if (!card) continue;
+        const w = p.params.widthDp;
+        const h = p.params.heightDp;
+        if (w && h && maxW > 0) {
+            card.style.setProperty("--size-ratio", (w / maxW).toFixed(4));
+            card.style.setProperty("--aspect-ratio", w + " / " + h);
+        } else {
+            card.style.removeProperty("--size-ratio");
+            card.style.removeProperty("--aspect-ratio");
+        }
+    }
 }
